# services/order/docs/order.openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: |
    REST API сервиса заказов (Tier C) на Node.js. Покрывает спеку `docs/spec/order-service.md`
    в объёме одиннадцатого шага практикума: идемпотентное создание черновика, чтение и переходы статусов.
    Ошибки - RFC 9457 Problem Details (`application/problem+json`), `type` вида `urn:problem:order:<CODE>`.
  version: 0.11.0

servers:
  - url: http://localhost:3081
    description: Local

security:
  - bearerAuth: []

paths:
  /api/v1/orders:
    post:
      operationId: createOrder
      summary: Создать черновик заказа
      description: |
        Цены позиций берутся из Catalog Service синхронно. Один заказ - товары одного продавца (BR-014).
        Если каталог недоступен, заказ не создаётся и возвращается `503 SERVICE_DEGRADED`.
        Повтор с тем же `Idempotency-Key` и тем же телом отдаёт прежний заказ ответом 200 (BR-010).
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Ключ идемпотентности от клиента, до 128 символов; обычно UUID
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Черновик создан
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '200':
          description: Повтор запроса с тем же Idempotency-Key - прежний заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: VALIDATION_ERROR (в том числе без Idempotency-Key), MALFORMED_REQUEST, EMPTY_ORDER, MULTI_SELLER_NOT_SUPPORTED
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          description: TOKEN_MISSING, TOKEN_INVALID
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: ACCESS_DENIED
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: PRODUCT_NOT_FOUND - каталог не нашёл товар
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: IDEMPOTENCY_KEY_CONFLICT - тот же ключ с другим телом
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: SERVICE_DEGRADED - каталог недоступен, заказ не создан
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/v1/orders/{orderId}:
    get:
      operationId: getOrder
      summary: Получить заказ
      description: Владелец видит свой заказ, администратор - любой; чужой заказ отвечает 404.
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: ORDER_NOT_FOUND
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /api/v1/orders/{orderId}/confirm:
    post:
      operationId: confirmOrder
      summary: Подтвердить черновик (DRAFT -> PENDING_PAYMENT)
      description: Покупатель или администратор. Сумма не меньше 100 рублей, иначе ORDER_BELOW_MINIMUM.
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Заказ, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '400': { description: ORDER_BELOW_MINIMUM, EMPTY_ORDER, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '409': { description: ORDER_INVALID_STATE, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      summary: Отменить заказ
      description: |
        Из DRAFT и PENDING_PAYMENT - без возврата. Из PAID - сага: возврат в сервисе платежей, потом CANCELLED.
        Платежи недоступны - 503 SERVICE_DEGRADED, заказ остаётся PAID.
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reasonCode]
              properties:
                reasonCode: { type: string, example: CHANGED_MIND }
                comment: { type: string, maxLength: 500 }
      responses:
        '200': { description: Заказ, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '409': { description: ORDER_INVALID_STATE, REFUND_REJECTED, PAYMENT_NOT_FOUND, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '503': { description: SERVICE_DEGRADED, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/orders/{orderId}/ship:
    post:
      operationId: shipOrder
      summary: Передать в доставку (PAID -> SHIPPED), продавец заказа
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [trackingNumber], properties: { trackingNumber: { type: string } } }
      responses:
        '200': { description: Заказ, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '404': { description: ORDER_NOT_FOUND - чужой заказ, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '409': { description: ORDER_INVALID_STATE, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/orders/{orderId}/deliver:
    post:
      operationId: confirmDelivery
      summary: Подтвердить получение (SHIPPED -> DELIVERED), покупатель
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Заказ, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '409': { description: ORDER_INVALID_STATE, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/orders/{orderId}/pay:
    post:
      operationId: payOrder
      summary: Отметить оплату (PENDING_PAYMENT -> PAID), только администратор
      description: Для стенда. В бою оплату приносит событие PaymentCompleted из marketplace.payments.v1.
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [paymentId], properties: { paymentId: { type: string, format: uuid } } }
      responses:
        '200': { description: Заказ, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '409': { description: ORDER_INVALID_STATE, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Address:
      type: object
      required: [country, city, street, postalCode]
      properties:
        country: { type: string, example: RU }
        city: { type: string, example: Москва }
        street: { type: string, example: Тверская, 1 }
        postalCode: { type: string, example: '125009' }
        pickupPoint: { type: string }

    OrderItemRequest:
      type: object
      required: [productId, sellerId, quantity]
      properties:
        productId: { type: string, format: uuid }
        sellerId: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 999 }

    CreateOrderRequest:
      type: object
      required: [items, shippingAddress]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderItemRequest' }
        shippingAddress: { $ref: '#/components/schemas/Address' }

    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        sellerId: { type: string, format: uuid }
        quantity: { type: integer }
        unitPrice: { type: number, example: 2490.5 }
        lineTotal: { type: number, example: 4981 }

    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        sellerId: { type: string, format: uuid }
        status:
          type: string
          enum: [DRAFT, PENDING_PAYMENT, PAID, SHIPPED, DELIVERED, COMPLETED, EXPIRED, CANCELLED, DISPUTE, REFUNDED]
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        shippingFee: { type: number }
        total: { type: number }
        currency: { type: string, example: RUB }
        shippingAddress: { $ref: '#/components/schemas/Address' }
        paymentId: { type: string, format: uuid }
        paidAt: { type: string, format: date-time }
        shippedAt: { type: string, format: date-time }
        deliveredAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Problem:
      type: object
      properties:
        type: { type: string, example: 'urn:problem:order:SERVICE_DEGRADED' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        errors:
          type: object
          additionalProperties: { type: string }